refactor: theme drift-guard uses a package-private accessor instead of setAccessible (Codacy)
Moves ReportThemeConsistencyTest into AllureManager's package and reads the injected theme CSS via a new package-private allureThemeColorsStyle() accessor, dropping the reflection setAccessible() Codacy flagged.
