+++ shaft-engine/src/main/java/com/shaft/tools/io/internal/AllureManager.java
@@ -297,6 +297,16 @@ public class AllureManager {
             </style>
             """;
 
+    /**
+     * Exposes the injected Allure theme-color overrides for the brand-palette drift-guard test
+     * (issue #3504), so consistency with {@code ReportHtmlTheme} is checked without reflection.
+     *
+     * @return the {@code <style>} block SHAFT patches into the generated Allure report
+     */
+    static String allureThemeColorsStyle() {
+        return ALLURE_THEME_COLORS_STYLE;
+    }
+
     // ─── Portable Node.js bootstrap ────────────────────────────────────────────
     /** Cache directory for the downloaded portable Node.js distribution. */
     private static final String NODEJS_CACHE_DIR = System.getProperty("user.home")
